+++ src/my_aes.h
@@ -64,7 +64,7 @@ int my_aes256_decrypt(const char *source, int source_length, char *dest,
   returns  - size of buffer required to store encrypted data
 */
 
-int my_aes256_get_size(int source_length);
+unsigned long my_aes256_get_size(int source_length);
 
 C_MODE_END
